+++ api/src/controller/order.controller.js
@@ -4,7 +4,6 @@ import Cart from "../models/cart.model.js";
 import Order from "../models/order.model.js";
 import OrderItem from "../models/orderItems.js";
 import CartItem from "../models/cartItem.model.js";
-import { isPinCodeAvailable } from "../utils/isPincodeAvailable.js";
 
 const placeOrder = asyncHandler(async (req, res) => {
   const { id } = req.user;
@@ -126,36 +125,4 @@ const cancelOrder = asyncHandler(async (req, res) => {
     .json(new ApiResponse(200, "Order cancelled successfully", order));
 });
 
-const isPincodeAvailable = asyncHandler(async (req, res) => {
-  const { pincode } = req.body;
-  try {
-    const isAvailable = await isPinCodeAvailable(pincode);
-
-    if (!isAvailable) {
-      return res
-        .status(400)
-        .json(
-          new ApiResponse(400, "Delivery not available in your area", null)
-        );
-    }
-
-    return res
-      .status(200)
-      .json(
-        new ApiResponse(200, "Delivery available in your area", isAvailable)
-      );
-  } catch (error) {
-    console.error(error);
-    return res
-      .status(500)
-      .json(
-        new ApiResponse(
-          500,
-          "Error fetching pincode availability",
-          error.message
-        )
-      );
-  }
-});
-
-export { cancelOrder, placeOrder, getAllOrders, isPincodeAvailable };
+export { cancelOrder, placeOrder, getAllOrders };
