+++ api/src/routes/order.route.js
@@ -3,12 +3,10 @@ import {
   placeOrder,
   getAllOrders,
   cancelOrder,
-  isPincodeAvailable,
 } from "../controller/order.controller.js";
 import { verifyToken } from "../middleware/verifyUser.js";
 const OrderRouter = express.Router();
 OrderRouter.route("/placeOrder").post(verifyToken, placeOrder);
 OrderRouter.route("/cancelOrder").get(verifyToken, cancelOrder);
 OrderRouter.route("/getAllOrders").get(verifyToken, getAllOrders);
-OrderRouter.route("/isPincodeAvailable").post(isPincodeAvailable);
 export default OrderRouter;
